+++ CHANGELOG.md
@@ -16,6 +16,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1469 Update footer with "Accompanying documents" dropdown
 - YLP-1475 Implement the team details page for patients
 - YLP-1476 Remove care teams page
+- YLP-1500 Patient can join a team from header
 ### Engineering Use
 - YLP-1256 Split Profile component into smaller subcomponents
 - YLP-1319 Switch-roles dialogs CSS not consistent with other dialogs